# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "{{.Description}}",
        "title": "{{.Title}}",
        "contact": {},
        "license": {},
        "version": "{{.Version}}"
    },
    "host": "{{.Host}}",
    "basePath": "{{.BasePath}}",
    "paths": {
...
        "/v2/contract/call": {
            "post": {
                "description": "仅在本节节点evm副本上执行合约，不会广播交易，不对区块产生影响，不消耗gas。用户估算智能合约消耗的gas，以及用于调用合约只读方法。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "v2-contract"
                ],
                "summary": "只读调用合约（非离线签名，慎用）",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/bean.ContractCallTx"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/bean.EvmCallResult"
                        }
                    }
                }
            }
        },
        "/v2/contract/deploy": {
            "post": {
                "description": "部署合约（私钥交易，非离线签名，慎用）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "v2-contract"
                ],
                "summary": "部署合约（非离线签名，慎用）",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/bean.ContractDeployTx"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/bean.PublicResp"
                        }
                    }
                }
            }
        },
...
        "/v2/contract/invoke": {
            "post": {
                "description": "调用合约（私钥交易，非离线签名，慎用）,使用本接口调用合约只读方法会消耗gas。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "v2-contract"
                ],
                "summary": "调用合约（非离线签名，慎用）",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/bean.ContractInvokeTx"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/bean.PublicResp"
                        }
                    }
                }
            }
        },
        "/v2/contract/multisigTransactions": {
            "post": {
                "description": "发起evm多签交易，支持OLO转账、合约交易",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "v2-contract"
                ],
                "summary": "发起evm多签交易",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/bean.SignedMultisigEvmTx"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/bean.PublicResp"
                        }
                    }
                }
            }
        },
        "/v2/contract/query": {
            "post": {
                "description": "仅在本节节点evm副本上执行合约，不会广播交易，不对区块产生影响，不消耗gas。用户估算智能合约消耗的gas，以及用于调用合约只读方法。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "v2-contract"
                ],
                "summary": "只读调用合约",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/bean.SignedEvmTx"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/bean.EvmCallResult"
                        }
                    }
                }
            }
        },
        "/v2/contract/sign": {
            "post": {
                "description": "根据transaction参数进行签名",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "v2-contract"
                ],
                "summary": "根据transaction参数进行签名",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/bean.SignEvmTx"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/bean.PublicResp"
                        }
                    }
                }
            }
        },
        "/v2/contract/transactions": {
            "post": {
                "description": "发起evm交易，支持普通转账、合约交易，普通转账默认使用本接口",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "v2-contract"
                ],
                "summary": "发起evm交易",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/bean.SignedEvmTx"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/bean.PublicResp"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "bean.ContractCallTx": {
            "type": "object",
            "properties": {
                "contract": {
                    "description": "合约地址",
                    "type": "string"
                },
                "gasLimit": {
                    "description": "gas限额",
                    "type": "integer"
                },
                "gasPrice": {
                    "description": "gas价格",
                    "type": "string"
                },
                "payload": {
                    "description": "负载数据，abi.pack(function+参数) hex编码字符串",
                    "type": "string"
                },
                "privkey": {
                    "description": "交易发起者私钥",
                    "type": "string"
                },
                "sender": {
                    "description": "交易发起者公钥",
                    "type": "string"
                },
                "value": {
                    "description": "金额，通常为0",
                    "type": "string"
                }
            }
        },
        "bean.ContractDeployTx": {
            "type": "object",
            "properties": {
                "gasLimit": {
                    "description": "gas限额",
                    "type": "integer"
                },
                "gasPrice": {
                    "description": "gas价格",
                    "type": "string"
                },
                "memo": {
                    "description": "备注",
                    "type": "string"
                },
                "payload": {
                    "description": "合约部署字节码",
                    "type": "string"
                },
                "privkey": {
                    "description": "交易发起者私钥",
                    "type": "string"
                },
                "sender": {
                    "description": "交易发起者公钥",
                    "type": "string"
                },
                "value": {
                    "description": "交易金额，通常为0",
                    "type": "string"
                }
            }
        },
        "bean.ContractInvokeTx": {
            "type": "object",
            "properties": {
                "contract": {
                    "description": "合约地址",
                    "type": "string"
                },
                "gasLimit": {
                    "description": "gas限额",
                    "type": "integer"
                },
                "gasPrice": {
                    "description": "gas价格",
                    "type": "string"
                },
                "memo": {
                    "description": "备注",
                    "type": "string"
                },
                "payload": {
                    "description": "合约负载 abi.pack(function+参数)",
                    "type": "string"
                },
                "privkey": {
                    "description": "交易发起者私钥",
                    "type": "string"
                },
                "sender": {
                    "description": "交易发起者公钥",
                    "type": "string"
                },
                "value": {
                    "description": "交易金额",
                    "type": "string"
                }
            }
        },
...
        "bean.EvmCallResult": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "错误码",
                    "type": "integer"
                },
                "gasUsed": {
                    "description": "消耗的gas",
                    "type": "integer"
                },
                "msg": {
                    "description": "msg",
                    "type": "string"
                },
                "ret": {
                    "description": "返回数据的hex编码",
                    "type": "string"
                }
            }
        },
...
        "bean.PublicResp": {
            "type": "object",
            "properties": {
                "isSuccess": {
                    "description": "是否成功",
                    "type": "boolean"
                },
                "message": {
                    "description": "消息提醒",
                    "type": "string"
                },
                "result": {
                    "description": "数据对象",
                    "type": "object"
                }
            }
        },
...
        "bean.SignedEvmTx": {
            "type": "object",
            "properties": {
                "body": {
                    "type": "object",
                    "properties": {
                        "load": {
                            "description": "合约负载，普通原声币转账时为空",
                            "type": "string"
                        },
                        "memo": {
                            "description": "备注",
                            "type": "string"
                        },
                        "to": {
                            "description": "交易接受者地址或合约地址",
                            "type": "string"
                        },
                        "value": {
                            "description": "交易金额",
                            "type": "string"
                        }
                    }
                },
                "createdAt": {
                    "description": "时间戳，可选",
                    "type": "integer"
                },
                "gasLimit": {
                    "description": "gas限额",
                    "type": "integer"
                },
                "gasPrice": {
                    "description": "gas价格，最低为1",
                    "type": "string"
                },
                "mode": {
                    "description": "交易模式，默认为0；0-同步模式 1-全异步 2-半异步；如果tx执行时间较长、网络不稳定、出块慢，建议使用半异步模式。",
                    "type": "integer"
                },
                "nonce": {
                    "description": "交易发起者nonce",
                    "type": "integer"
                },
                "sender": {
                    "description": "交易发起者公钥",
                    "type": "string"
                },
                "signature": {
                    "description": "交易签名",
                    "type": "string"
                }
            }
        },
        "bean.SignedMultisigEvmTx": {
            "type": "object",
            "properties": {
                "deadline": {
                    "description": "有效截止时间",
                    "type": "integer"
                },
                "from": {
                    "description": "多签账户地址",
                    "type": "string"
                },
                "gasLimit": {
                    "description": "gas限额",
                    "type": "integer"
                },
                "gasPrice": {
                    "description": "gas价格，最低为1",
                    "type": "string"
                },
                "load": {
                    "description": "合约负载，普通原声币转账时为空",
                    "type": "string"
                },
                "memo": {
                    "description": "备注",
                    "type": "string"
                },
                "mode": {
                    "description": "交易模式，默认为0；0-同步模式 1-全异步 2-半异步；如果tx执行时间较长、网络不稳定、出块慢，建议使用半异步模式。",
                    "type": "integer"
                },
                "nonce": {
                    "description": "多签账户nonce",
                    "type": "integer"
                },
                "signature": {
                    "description": "交易签名",
                    "type": "object",
                    "$ref": "#/definitions/bean.MultiSignature"
                },
                "to": {
                    "description": "交易接受者地址或合约地址",
                    "type": "string"
                },
                "value": {
                    "description": "交易金额",
                    "type": "string"
                }
            }
        },
...
    }
}